moved web_content/removed into web_content_removed
