+++ test/integration/test_cases/test_cases.json
@@ -1326,142 +1326,142 @@
                 "7f0cd74"
             ]
         },
-        "web_content/removed/animated/composition_comparison_pie_coxcomb_column_2dis_2con": {
+        "web_content_removed/animated/composition_comparison_pie_coxcomb_column_2dis_2con": {
             "refs": [
                 "bccd4e0"
             ]
         },
-        "web_content/removed/animated/composition_comparison_waterfall_column_2dis_1con": {
+        "web_content_removed/animated/composition_comparison_waterfall_column_2dis_1con": {
             "refs": [
                 "3a1a33b"
             ]
         },
-        "web_content/removed/animated/composition_percentage_area_stream_3dis_1con": {
+        "web_content_removed/animated/composition_percentage_area_stream_3dis_1con": {
             "refs": [
                 "8d25829"
             ]
         },
-        "web_content/removed/animated/composition_percentage_column_3dis_1con": {
+        "web_content_removed/animated/composition_percentage_column_3dis_1con": {
             "refs": [
                 "ffc318d"
             ]
         },
-        "web_content/removed/animated/composition_percentage_column_stream_3dis_1con": {
+        "web_content_removed/animated/composition_percentage_column_stream_3dis_1con": {
             "refs": [
                 "8b41c0e"
             ]
         },
-        "web_content/removed/animated/distribution_relationship_dotplot_dotplot": {
+        "web_content_removed/animated/distribution_relationship_dotplot_dotplot": {
             "refs": [
                 "9bf8509"
             ]
         },
-        "web_content/removed/animated/drill_aggreg_improve_line": {
+        "web_content_removed/animated/drill_aggreg_improve_line": {
             "refs": [
                 "54718c2"
             ]
         },
-        "web_content/removed/animated/drilldown_aggregate_line": {
+        "web_content_removed/animated/drilldown_aggregate_line": {
             "refs": [
                 "47693ee"
             ]
         },
-        "web_content/removed/animated/merge_split_area_stream_3dis_1con": {
+        "web_content_removed/animated/merge_split_area_stream_3dis_1con": {
             "refs": [
                 "fc9d408"
             ]
         },
-        "web_content/removed/animated/merge_split_bar": {
+        "web_content_removed/animated/merge_split_bar": {
             "refs": [
                 "4bc0761"
             ]
         },
-        "web_content/removed/animated/merge_split_radial_stacked_rectangle_2dis_1con": {
+        "web_content_removed/animated/merge_split_radial_stacked_rectangle_2dis_1con": {
             "refs": [
                 "01dae3c"
             ]
         },
-        "web_content/removed/animated/orientation_circle": {
+        "web_content_removed/animated/orientation_circle": {
             "refs": [
                 "8ba457a"
             ]
         },
-        "web_content/removed/animated/orientation_dot_circle": {
+        "web_content_removed/animated/orientation_dot_circle": {
             "refs": [
                 "0a92164"
             ]
         },
-        "web_content/removed/animated/orientation_marimekko_rectangle_2dis_2con": {
+        "web_content_removed/animated/orientation_marimekko_rectangle_2dis_2con": {
             "refs": [
                 "ba4645a"
             ]
         },
-        "web_content/removed/animated/orientation_rectangle": {
+        "web_content_removed/animated/orientation_rectangle": {
             "refs": [
                 "24c160b"
             ]
         },
-        "web_content/removed/animated/pie_donut2_rectangle_1dis_1con": {
+        "web_content_removed/animated/pie_donut2_rectangle_1dis_1con": {
             "refs": [
                 "0554a81"
             ]
         },
-        "web_content/removed/animated/relationship_comparison_circle_2_bubble_plot": {
+        "web_content_removed/animated/relationship_comparison_circle_2_bubble_plot": {
             "refs": [
                 "e76689d"
             ]
         },
-        "web_content/removed/animated/relationship_total_bubble_plot_column": {
+        "web_content_removed/animated/relationship_total_bubble_plot_column": {
             "refs": [
                 "ae768b7"
             ]
         },
-        "web_content/removed/animated/stack_group_area_line": {
+        "web_content_removed/animated/stack_group_area_line": {
             "refs": [
                 "a9baae0"
             ]
         },
-        "web_content/removed/animated/stack_group_circle": {
+        "web_content_removed/animated/stack_group_circle": {
             "refs": [
                 "c593f63"
             ]
         },
-        "web_content/removed/animated/stack_group_treemap": {
+        "web_content_removed/animated/stack_group_treemap": {
             "refs": [
                 "839e607"
             ]
         },
-        "web_content/removed/animated/total_element_bubble_2_bar": {
+        "web_content_removed/animated/total_element_bubble_2_bar": {
             "refs": [
                 "4f0c0dd"
             ]
         },
-        "web_content/removed/animated/total_element_bubble_column": {
+        "web_content_removed/animated/total_element_bubble_column": {
             "refs": [
                 "de3e165"
             ]
         },
-        "web_content/removed/animated/treemap_radial": {
+        "web_content_removed/animated/treemap_radial": {
             "refs": [
                 "1913066"
             ]
         },
-        "web_content/removed/animated/zoom_area": {
+        "web_content_removed/animated/zoom_area": {
             "refs": [
                 "e7ad338"
             ]
         },
-        "web_content/removed/animated/zoom_area_polar": {
+        "web_content_removed/animated/zoom_area_polar": {
             "refs": [
                 "f415237"
             ]
         },
-        "web_content/removed/animated/zoom_line": {
+        "web_content_removed/animated/zoom_line": {
             "refs": [
                 "7029760"
             ]
         },
-        "web_content/removed/animated/zoom_line_polar": {
+        "web_content_removed/animated/zoom_line_polar": {
             "refs": [
                 "e104842"
             ]
