+++ test/integration/test_cases/web_content_removed/animated/composition_comparison_pie_coxcomb_column_2dis_2con.mjs
@@ -1,4 +1,4 @@
-import { data } from "../../../../test_data/infinite_data.mjs";
+import { data } from "../../../test_data/infinite_data.mjs";
 
 const testSteps = [
   (chart) =>
