+++ test/integration/test_cases/web_content_removed/animated/composition_percentage_column_stream_3dis_1con.mjs
@@ -1,4 +1,4 @@
-import { data_14 } from "../../../../test_data/chart_types_eu.mjs";
+import { data_14 } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
