+++ test/integration/test_cases/web_content_removed/animated/orientation_marimekko_rectangle_2dis_2con.mjs
@@ -1,4 +1,4 @@
-import { data_4 } from "../../../../test_data/chart_types_eu.mjs";
+import { data_4 } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
