+++ test/integration/test_cases/web_content_removed/animated/relationship_comparison_circle_2_bubble_plot.mjs
@@ -1,4 +1,4 @@
-import { data } from "../../../../test_data/chart_types_eu.mjs";
+import { data } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
