+++ test/integration/test_cases/web_content_removed/animated/stack_group_circle.mjs
@@ -1,4 +1,4 @@
-import { data } from "../../../../test_data/chart_types_eu.mjs";
+import { data } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
