+++ test/integration/test_cases/web_content_removed/animated/total_element_bubble_2_bar.mjs
@@ -1,4 +1,4 @@
-import { data } from "../../../../test_data/chart_types_eu.mjs";
+import { data } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
@@ -6,13 +6,15 @@ const testSteps = [
       data: data,
       config: {
         channels: {
-          x: ["Joy factors", "Value 6 (+/-)"],
-          y: "Value 5 (+/-)",
           color: "Joy factors",
-          size: "Value 2 (+)",
           label: "Country_code",
+          /* With a dimension on the size channel Vizzu will stack
+                the elements by the categories on the other channels
+                on charts without coordinates. Here the Country code dimension is
+                used to stack the bubbles by the dimension on the color channel. */
+          size: ["Country_code", "Value 2 (+)"],
         },
-        title: "Bubble Plot",
+        title: "Stacked Bubble Chart",
         geometry: "circle",
       },
     }),
@@ -21,9 +23,12 @@ const testSteps = [
     chart.animate({
       config: {
         channels: {
+          x: ["Country_code", "Value 2 (+)"],
           y: "Joy factors",
-          x: ["Value 2 (+)", "Country_code"],
           label: null,
+          /* The stacking is eliminated when we remove
+                the extra dimension from the size channel. */
+          size: null,
         },
         title: "Bar Chart",
         geometry: "rectangle",
