+++ test/integration/test_cases/web_content_removed/animated/zoom_area.mjs
@@ -1,4 +1,4 @@
-import { data_6 } from "../../../../test_data/chart_types_eu.mjs";
+import { data_6 } from "../../../test_data/chart_types_eu.mjs";
 
 const testSteps = [
   (chart) =>
@@ -7,22 +7,21 @@ const testSteps = [
       config: {
         channels: {
           x: ["Year", "Joy factors"],
-          y: "Value 3 (+)",
+          y: ["Value 3 (+)", "Country_code"],
           color: "Country_code",
         },
-        title: "Polar Line chart",
-        geometry: "line",
-        coordSystem: "polar",
+        title: "Stacked Area",
+        geometry: "area",
       },
     }),
   (chart) =>
     chart.animate({
       data: {
         filter: (record) =>
-          data_6.filter(record) && record.Year < 8 && record.Year > 2,
+          data_6.filter(record) && record.Year < 12 && record.Year > 6,
       },
       config: {
-        title: "Zoomed Polar Line chart",
+        title: "Zoomed Stacked Area",
       },
     }),
 ];
